# alodo-backend/public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "ALODO — API du diagnostic MPME",
        "description": "Documentation générée depuis les annotations des contrôleurs. Obtenir data.token via login/register, puis utiliser Authorize. Les identifiants des exemples sont fictifs.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Backend courant"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Créer un compte entreprise",
                "description": "AuthController::register. Crée utilisateur, diagnostic baseline et token dans une transaction. Le rôle user est attribué par le serveur. Aucun paramètre de rôle accepté.",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AuthData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/Validation"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": []
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se connecter",
                "description": "AuthController::login. Une connexion réussie supprime tous les anciens tokens du compte avant de créer le nouveau.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AuthData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/Validation"
                    }
                },
                "security": []
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Se déconnecter",
                "description": "AuthController::logout. Révoque tous les tokens du compte, pas uniquement le token courant.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "string",
                                            "enum": [
                                                null
                                            ],
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Lire le compte connecté",
                "description": "Closure routes/api.php. Réponse sans enveloppe success/message/data. La relation role n’est pas chargée.",
                "operationId": "currentUser",
                "responses": {
                    "200": {
                        "description": "Utilisateur directement, sans enveloppe.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Role": {
                "required": [
                    "id",
                    "intitule"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "intitule": {
                        "type": "string",
                        "enum": [
                            "user",
                            "admin"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "User": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "role_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "role_id": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AuthenticatedUser": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/User"
                    },
                    {
                        "required": [
                            "role"
                        ],
                        "properties": {
                            "role": {
                                "$ref": "#/components/schemas/Role"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "AuthData": {
                "required": [
                    "token",
                    "user"
                ],
                "properties": {
                    "token": {
                        "description": "Token Sanctum à copier dans Authorize, sans le préfixe Bearer.",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/AuthenticatedUser"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1,
                        "example": "Atelier Exemple"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "entreprise@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "example": "ExempleMotDePasse2026!"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "entreprise@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "example": "ExempleMotDePasse2026!"
                    }
                },
                "type": "object"
            },
            "Error": {
                "description": "Les erreurs Laravel peuvent ne contenir que message. Les contrôleurs peuvent ajouter success:false et errors:null.",
                "required": [
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            false
                        ]
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            false
                        ]
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object",
                "example": {
                    "message": "The given data was invalid.",
                    "errors": {
                        "responses.0.valeur": [
                            "The selected responses.0.valeur is invalid."
                        ]
                    }
                }
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "401 : token absent/invalide ou identifiants de connexion incorrects.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "404 : ressource absente, diagnostic inaccessible ou rôle non configuré.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Validation": {
                "description": "422 : données invalides. Les clés responses.{index}.valeur concernent la soumission ; answers.{question_id} concerne la complétude finale. diagnostic_id peut signaler un diagnostic déjà terminé.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Laravel Sanctum, sans préfixe Bearer dans Authorize.",
                "scheme": "bearer"
            }
        }
    },
...
}